fix(eslint): separate formSchema into a different file
